Refactor game design
Splitting game logic into several files is a good
practice.
