+++ devolution-web/src/game/design/income-action.ts
@@ -1,13 +1,5 @@
 import Decimal from 'break_infinity.js';
 
-export interface ClickAction {
-  id: string;
-  name: string;
-  description: string;
-  minCoinsGained: Decimal;
-  maxCoinsGained: Decimal;
-}
-
 export interface IncomeAction {
   id: string;
   name: string;
@@ -19,16 +11,6 @@ export interface IncomeAction {
     exponent: Decimal;
   };
 }
-
-export const BASIC_CLICK_ACTIONS: ClickAction[] = [
-  {
-    id: 'basic-click-action-1',
-    name: 'Coding',
-    description: 'Start writing your first lines of code',
-    minCoinsGained: Decimal.fromNumber(4),
-    maxCoinsGained: Decimal.fromNumber(8),
-  },
-];
 
 export const BASIC_INCOME_ACTIONS: IncomeAction[] = [
   {
