+++ devolution-web/src/game/design/index.ts
@@ -1,0 +1,2 @@
+export * from './click-action';
+export * from './income-action';
